Bring back the new-* template tests
