+++ examples/new-esm/template.spec.mjs
@@ -1,9 +1,15 @@
 import path from 'path'
+import {fileURLToPath} from 'url'
 import {execSync} from 'child_process'
-import {extensionFixtures, takeScreenshot} from '../extension-fixtures'
+import {extensionFixtures, takeScreenshot} from '../extension-fixtures.mjs'
 
-const exampleDir = 'examples/new-config-babel'
+// Recreate __dirname for ES Modules
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = path.dirname(__filename)
+
+const exampleDir = 'examples/new-esm'
 const pathToExtension = path.join(__dirname, `dist/chrome`)
+
 const test = extensionFixtures(pathToExtension, true)
 
 test.beforeAll(async () => {
@@ -17,7 +23,10 @@ test('should exist an element with the welcome message text', async ({
 }) => {
   await page.goto('chrome://newtab/')
   const h1 = page.locator('h1')
-  await test.expect(h1).toContainText('Welcome to your')
+    test.expect(h1).toContainText('Welcome to your');
+
+  // Take a screenshot once the element is found
+  await takeScreenshot(page, path.join(__dirname, 'screenshot.png'));
 })
 
 test('should exist a default color value', async ({page}) => {
@@ -25,15 +34,9 @@ test('should exist a default color value', async ({page}) => {
   const h1 = page.locator('h1')
   const color = await page.evaluate(
     (locator) => {
-      return window.getComputedStyle(locator!).getPropertyValue('color')
+      return window.getComputedStyle(locator).getPropertyValue('color')
     },
     await h1.elementHandle()
   )
-  await test.expect(color).toEqual('rgb(201, 201, 201)')
-})
-
-test.skip('takes a screenshot of the page', async ({page}) => {
-  await page.goto('chrome://newtab/')
-  await page.waitForSelector('h1')
-  await takeScreenshot(page, path.join(__dirname, 'screenshot.png'))
+  test.expect(color).toEqual('rgb(201, 201, 201)')
 })
