+++ examples/new-node-apis/template.spec.ts
@@ -2,7 +2,7 @@ import path from 'path'
 import {execSync} from 'child_process'
 import {extensionFixtures, takeScreenshot} from '../extension-fixtures'
 
-const exampleDir = 'examples/new-less'
+const exampleDir = 'examples/new-node-apis'
 const pathToExtension = path.join(__dirname, `dist/chrome`)
 const test = extensionFixtures(pathToExtension, true)
 
@@ -17,7 +17,7 @@ test('should exist an element with the welcome message text', async ({
 }) => {
   await page.goto('chrome://newtab/')
   const h1 = page.locator('h1')
-  await test.expect(h1).toContainText('Welcome to your')
+  test.expect(h1).toContainText('Welcome to your')
 })
 
 test('should exist a default color value', async ({page}) => {
@@ -29,7 +29,7 @@ test('should exist a default color value', async ({page}) => {
     },
     await h1.elementHandle()
   )
-  await test.expect(color).toEqual('rgb(201, 201, 201)')
+  test.expect(color).toEqual('rgb(201, 201, 201)')
 })
 
 test.skip('takes a screenshot of the page', async ({page}) => {
